+++ R/actions.R
@@ -255,10 +255,10 @@ gui_update_filter_class <- function(step_id, filter_id, show, class, session) {
 
 update_filter_gui <- function(cohort, step_id, filter_id, update, reset, session) {
   filter <- cohort$get_filter(step_id, filter_id)
-  run_on_request <- cohort$attributes$run_button
   updated_input <- FALSE
   updated_plot <- FALSE
-  if (("post_input" %in% update) && filter$gui$post_stats) {
+
+  if (("post_input" %in% update) && !identical(filter$gui$post_stats, FALSE)) {
     update <- c(update, "input")
   }
   if (("multi_input" %in% update) && filter$gui$multi_input) {
@@ -281,7 +281,11 @@ update_filter_gui <- function(cohort, step_id, filter_id, update, reset, session
       show <- FALSE
     }
     gui_update_filter_class(step_id, filter_id, show, "cb_no_data", session)
-    if (cohort$attributes$feedback) {
+    show_feedback <- if_null_default(
+      filter$get_params("feedback"),
+      cohort$attributes$feedback
+    )
+    if (show_feedback) {
       updated_plot <- TRUE
       gui_update_plot(step_id, filter_id, cohort, session) # todo optmize to not extract filter again inside plot
     }
@@ -361,10 +365,15 @@ update_next_step <- function(cohort, step_id, reset, session) {
 }
 
 input_val_handler <- function(val) {
-  if (is.list(val) && is.null(names(val))) {
-    return(unlist(val, recursive = TRUE))
-  }
-  else {
+  if (is.list(val)) {
+    if (is.null(names(val))) {
+      return(unlist(val, recursive = TRUE))
+    }
+    return(
+      val %>% purrr::map(unlist) %>%
+        purrr::map_if(~all(.x %in% c("TRUE", "FALSE", "NA")), as.logical)
+    )
+  } else {
     return(val)
   }
 }
@@ -386,11 +395,13 @@ convert_input_value <- function(changed_input, step_id, filter_id, cohort, updat
 
 gui_update_filter <- function(cohort, changed_input, session) {
 
-  run_on_request <- cohort$attributes$run_button
+  run_on_request <- !is_none(cohort$attributes$run_button)
   update_active <- changed_input$input_name == "active"
 
   step_id <- changed_input$step_id
   filter_id <- changed_input$filter_id
+  data_filter <- cohort$get_filter(step_id, filter_id)
+
   force_render <- getOption("scb_render_all", default = FALSE)
 
   print_state("update_filter", changed_input)
@@ -413,25 +424,29 @@ gui_update_filter <- function(cohort, changed_input, session) {
   if (!force_render && !is.null(changed_input$active)) {
     run_update <- !insert_filter(step_id, filter_id, cohort, session)
   }
+  filter_stats <- if_null_default(
+    data_filter$get_params("stats"),
+    cohort$attributes$stats
+  )
+  post_stats_visible <- "post" %in% filter_stats
   if (run_update) {
     update <- c("plot", "multi_input")
-    if (!run_on_request) {
+    if (!run_on_request && post_stats_visible) {
       update <- c(update, "post_input")
     }
     update_filter_gui(cohort, step_id, filter_id, update, FALSE, session)
   }
 
-  post_stats_visible <- "post" %in% cohort$attributes$stats
-  if (!run_on_request && post_stats_visible) {
+  if (!run_on_request && ("post" %in% cohort$attributes$stats)) {
     update <- "post_input"
-    gui_update_filters_loop(cohort, step_id, update, FALSE, exclude = filter_id, session)
+    gui_update_filters_loop(cohort, step_id, FALSE, update, exclude = filter_id, session)
   }
 
   if (update_active) {
     gui_update_filter_class(step_id, filter_id, changed_input$active, "hidden-input", session)
   }
 
-  if (!cohort$attributes$run_button) {
+  if (is_none(cohort$attributes$run_button)) {
     gui_update_data_stats(cohort, list(step_id = step_id), session)
     update_next_step(cohort, step_id, FALSE, session)
   }
@@ -733,7 +748,7 @@ gui_show_repro_code <- function(cohort, changed_input, session) {
     title = "Reproducible code",
     shiny::tags$code(
       class = "hl background",
-      cohort$get_code(width = I(80), output = FALSE)$text.tidy %>%
+      cohort$get_code(width = I(120), output = FALSE)$text.tidy %>%
         highr::hi_html() %>%
         purrr::map_chr(add_trailing_space) %>%
         paste(collapse = "</br>") %>%
@@ -1082,7 +1097,7 @@ gui_clear_step <- function(cohort, changed_input, session) {
 
   reset_filters(cohort, changed_input$step_id)
 
-  if (cohort$attributes$run_button) {
+  if (!is_none(cohort$attributes$run_button)) {
     trigger_pending_state(changed_input$step_id, "add", session)
     changed_input$run_flow <- FALSE
   }