+++ NEWS.md
@@ -1,3 +1,10 @@
-# cohortBuilder 0.1
+# shinyCohortBuilder (development version)
+
+* Arguments `feedback` and `stats` passed to filter definitions are now respected and override the arguments passed to `cb_server`.
+* Fix updating filters data when other filters are changed (https://github.com/r-world-devs/shinyCohortBuilder/issues/27).
+* Implement bugs related to `multi_discrete` filter (https://github.com/r-world-devs/shinyCohortBuilder/issues/9).
+* `onUnload` hook implemented to clear custom package hooks (https://github.com/r-world-devs/shinyCohortBuilder/issues/21).
+
+# shinyCohortBuilder 0.1
 
 * First release.